[Driver][SYCL] Fix issue using -fsycl-force-target with offload dependencies (#7025)
When using performing offloading with archives, an additional
clang-offload-deps step is taken that determines the device dependencies
during the link. When using -fsycl-force-target, we need to be sure that
the target in which the dependency information is associated with
matches so the proper device objects are referenced and pulled out.