[SYCL] Add ext_intel_device_id to sycl_ext_intel_device_info (#7010)
This adds the ability to query the device ID of a device, which must be
equal to the PCI device ID for PCI devices. An important use case of
this feature is detecting the GPU architecture, which is necessary in
applications that have dedicated code paths for different architectures.
Querying the device ID is implemented for Level Zero and OpenCL as well.