[cmake] validate `CHECK_SYCL_ALL` more thoroughly during configuration (intel/llvm-test-suite#1321)
I had a bit of trouble working out why lit was failing to start after
configuration had succceeded. Turns out I was passing a partial tuple
to cmake `-DCHECK_SYCL_ALL` due to a typo in a shell variable.

This patch verifies that both parts of each tuple in the
`CHECK_SYCL_ALL` variable are nonempty and terminates configuration with
an error message otherwise, which should make it a bit more obvious
what's happening.