[SYCL][NFC] Fix bugs in unit tests
* Define copy/move constructors and copy/move assignment operators for
PIImage class. This is required because we need to explicitly update
pointers in the MBinaryDesc of this class. It's needed even for the
move constructor and assignment operator because "short string optimization"
may cause addresses to change. Using constructors generated by default
leads to dereferencing invalid pointers and flaky failures.
* Remove cend() iterator dereference in unit tests.
* Use different kernel identifiers for different tests.All device images
are stored in the same program manager object and tests doesn't take that
into account. So, using the same kernel identifier in different tests
leads to multiple device images (related to different tests) in the program
manager associated with the same kernel identifier. It leads to the situation
when test's behaviour is affected by artifacts from another test.
* Fixed redefined compile/link/build plugin functions. Currently they may
produce a result with redundant space symbol.