[SYCL] Stop reinterpret from changing buffer_allocator type to const (#6769)
This PR prevents the `buffer_allocator` from being rebound to a const
type. Currently when a `buffer` is reinterpreted from type `T` to type
`const T` the `buffer_allocator` type is also changed to `const T`. This
does not agree with the SYCL spec which states "A buffer of data type
const T uses buffer_allocator<T> by default."

This PR also adds a compile time test which validates the returned type.