[SYCL] Wait for stream service tasks
Before the patch explicit waiting for the task which is expected to
print a message using sycl::stream doesn't guarantee that the message is
actually printed(the printing may happen later).
The patch adds such a guarantee by regestering events of stream service
tasks in the user event produced for the original user task as well as
in the queue where this task is submitted to. Then on explicit calls to
queue::wait and event::wait we make sure that these additional events
are complete as well.