[SYCL][Docs] Mark host as deprecated in SYCL_DEVICE_FILTER
The host device has been removed and related APIs have been deprecated.
This includes the `host` device-type and backend in
`SYCL_DEVICE_FILTER`. This commit changes the documentation for
`SYCL_DEVICE_FILTER` to mention the deprecation of `host` in the filter
and remove any mention of the behavior of using the deprecated
values and interfaces.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>