[SYCL] Native event for default-ctored sycl::event has to be in COMPLETE state
Per SYCL 2020 for event():

   > Constructs an event that is immediately ready. The event has no
   > dependencies and no associated commands. Waiting on this event will
   > return immediately and querying its status will return
   > info::event_command_status::complete.

piEventCreate on the other hand creates an event that isn't completed.
As such an extra call to piEventSetStatus is needed making that API
not-deprecated.

There is a more general problem that isn't addressed here:

  auto e = q.submit(... h.host_task(...) ..)

This event would be a host one and we assert that no get_native could be
called on it (see existing sycl::detail::getImplBackend). If we will ever
want to support such scenario we'd need to implement some tracking of
host/backed events in the SYCL RT and keep updating the latter whenever
the host one changes the state.