[SYCL] Select a more consistent image ID for caching
The program cache was recently changed to use the device image pointer
as part of its cache key. However, pointers are unreliable as assigning
device images may mean that device images may change identifiers. This
commit changes the selected identifier to be a counter instead of using
the pointer value.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>