[SYCL] Select a more consistent image ID for caching (#7154)
The program cache was recently changed to use the device image pointer
as part of its cache key. However, the `RTDeviceBinaryImage` class has
move constructor and assignment operator defined to allow them to
transition between collections, so the pointers can change, making them
an unreliable cache key. This commit changes the selected identifier to
be the pointer to the underlying binary image instead.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>