[SYCL][Docs] Move sycl_ext_oneapi_kernel_properties to experimental
This commit moves the sycl_ext_oneapi_kernel_properties extension from
proposal to experimental. As part of the move, the extension is
transitioned to the new extension template.

Additionally it adds a note about the current order limitation of
`SYCL_EXT_ONEAPI_FUNCTION_PROPERTY` and `SYCL_EXTERNAL` in the current
implementation.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>