[SYCL] Add property-based device_has diagnostics
The device_has property should have behavior similar to that of the
sycl::device_has attribute. This commit makes the aspect propagation
pass issue the same diagnostic as done for when an application uses
aspects that are not in sycl::device_has, albeit with a small change to
the diagnostics message to differentiate the origin of the device_has,
be it property or attribute.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>