[SYCL] Fix use-after-move bug in building for multiple devices
The cache key used while building kernel binaries will have the spec
constant data moved to it to avoid repeated copies of a potentially
large data structure. However, when building for multiple devices the
spec constant data is moved each time a the cache key is created, which
can corrupt the cache key. This commit fixes this by creating the cache
key once and mutating it for each insertion, preserving common parts of
they key and preventing use-after-move.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>