[Driver][SYCL] Allow for -fcf-protection to run on host only
The -fcf-protection option is not supported for spir64 targets.  This causes
an error during device compilation when compiled with -fsycl.  Allow the
option to be accepted and work on the host compilation side only, emitting
a diagnostic that the option is being ignored for the target compiles.

before: clang++ -fsycl -fcf-protection -c a.cpp
error: option 'cf-protection=return' cannot be specified on this target
error: option 'cf-protection=branch' cannot be specified on this target
2 errors generated.

after: clang++ -fsycl -fcf-protection -c a.cpp
clang-16: warning: ignoring '-fcf-protection' option as it is not currently supported for target 'spir64-unknown-unknown' [-Woption-ignored]