[SYCL] Rename DoubleGRF to LargeGRF (#7284)
This change renames double GRF to large GRF both for users and
internally in the compiler. We're doing this because we got direct
feedback from customer facing engineers that we should use the large GRF
terminology, and it also makes the naming consistent with other compiler
work we are doing.

For the user,

```
set_kernel_properties(kernel_properties::use_double_grf);
```
still works, it will just throw a deprecated warning and will be removed
in a future release.

The new way is
```
set_kernel_properties(kernel_properties::use_large_grf);
```

There should be no ABI break because we still check the previous image
property name in the program manager, so applications built with an old
compiler work using the runtime from a new compiler. I confirmed this
with manual testing.

I will update the system test here to test the new flag as well:
https://github.com/intel/llvm-test-suite/blob/intel/SYCL/DeviceCodeSplit/double-grf.cpp

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>