[SYCL] missing SYCL_EXPORT leading to downstream troubles (#7318)
my earlier post-commit fix caused an ABI symbols problem downstream. But
the original problem is that the custom specialization of
`get_info<parent_device>` is merely missing an export declaration. If
I'm right, or just dumb lucky, then this will fix all downstream
failures, including post-commit. And if not, well, the symbol update is
correct regardless

Signed-off-by: Chris Perkins <chris.perkins@intel.com>