[Driver][SYCL] Update archive creation with -fsycl-link for FPGA
When generating archives using -fsycl-link=early|image, any associated
generated objects and device binaries are added to the archive for later
processing.  Archives can also be used in this manner, but creating the
archives with 'cr' would just add the archive to the final generated
archive, which does not allow the final archive to be used for host linking.

Update the archive generation by using 'cqL' for llvm-ar, as that allows
for any archives that are pulled in to be added via the contents instead of
the archives themselves.