[NFC][SYCL][Reduction] Rethink getReadWriteAccessorToInitializedMem helper
Previously the helper abstraction consisted of two implicitly tied
pieces that had to be used together and used member variable MOutBufPtr
to pass some data behind the scene. Change that to a new abstraction
withInitializedMem accepting a callable argument to make the data flow
explicit and more immediately clear.

This change also allows to change/optimize this piece in one place
without the need to modify multiple uses.