[SYCL] Remove use of SYCL_DEVICE_FILTER from set-arg-local-accessor (intel/llvm-test-suite#1375)
The set-arg-local-accessor test uses SYCL_DEVICE_FILTER explicitly
which causes failures with recent change to use ONEAPI_DEVICE_SELECTOR
because they are mutually exclusive. Since ONEAPI_DEVICE_SELECTOR is
added through the *_RUN_PLACEHOLDER macros the SYCL_DEVICE_FILTER is not
needed. The filtering for the OpenCL backend is ensured by the `opencl`
feature requirement at the top of the test.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>