[SYCL] Support negative filters for ONEAPI_DEVICE_SELECTOR (#7309)
This PR aims to add support for negative filters for the
ONEAPI_DEVICE_SELECTOR variable to provide the user with a more flexible
way of specifying which devices should and should not be available for
usage. For example, ONEAPI_DEVICE_SELECTOR='opencl:*;**!opencl:gpu**'
considers all opencl backend devices except for those that are of the
gpu type.