[CI][2/3] Add Ubuntu 22 containers build
Follow up on #7106, this is patch 2 of 3 patches expected:

Start running nightly on 22.04 containers in addition to 20.04 ones and push
nightly containers.
This is basically a copy-paste from 20.04 parts.

As before push is not expected to happen until this change is actually
merged. Based on the previous patch testing it seems it works well after
merge and doesn't really work in pull_request mode.

Adding more workflows to ignore in sycl_precommit.