Add a unified getSPIRVType interface. (#1716)
We have a couple of interfaces that create the custom SPIR-V types. Again, in
preparation for an eventual OpaqueType-based interface for types, these
functions are being replaced with a more centralized, uniform basis for
constructing these types.

Some of the related helper methods that were previously used to build the names
for struct types are now obsoleted by this change and have been removed.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/8784853