[SYCL] Remove use of potentially invalid parameter from aspect queries
Select device aspect queries were using an output type for extension
device queries that did not match the output type of the corresponding
info query. This commit changes these to not specify the size of the
output but instead base the aspect checks on whether the backend is able
to return a size of the output.

Fixes https://github.com/intel/llvm/issues/7011.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>