[SYCL] Move offload mismatch warning into SYCLActionBuilder
The SYCLActionBuilder handles the parsing of the arguments to determine
the SYCL triples and architectures being used, so doing the check here
means we can just re-use that rather than having to parse the triples
and target arguments again.

This fixes the issue where for CUDA and HIP the specific architecture is
embedded in the offload binary section name but is passed as a different
argument and is not in the triple of `-fsycl-targets`.