[SYCL] Change sycl::reqd_work_group_size with optional dimensions
In the current implementation the sycl::reqd_work_group_size attribute
sets the Y and Z dimension arguments optional. However, when the
internal representation of the attribute is created it will be padded
with 1's in the additional dimensions. An effect of this padding is that
dimensionality information is lost, which has three big drawbacks:
 1. SYCL work-group sizes are reversed, but only the specified
    dimensions are reversed. For example
    `sycl::reqd_work_group_size(1, 2, 3)` is for the backends the same
    as a backend work-group of size `<3,2,1>`, but
    `sycl::reqd_work_group_size(3)` corresonds to a backend work-group
    of `<3,1,1>` rather than `<1,1,3>`.
 2. The SYCL runtime is supposed to throw an exception when a kernel
    is launched with a number of dimensions that does not match the
    `sycl::reqd_work_group_size`'s dimensionality. sycl-post-link
    generates kernel meta-information for the runtime which could be
    used to diagnose these, but since the attribute is padded with 1's
    the `reqd_work_group_size` metadata node knows no difference between
    it and a user-specified attribute with trailing 1's.
 3. Sema cannot know the difference between two attributes where one was
    padded with 1's by the user and one was not, so it currently thinks
    these are equivalent.

To fix these, this patch changes Sema to not add the padding and instead
consider cases where the Y and Z dimensions are unset. This only affects
the SYCL spelling of the attribute.
Additionally, when generating the `reqd_work_group_size` attribute
CodeGen will only generate a metadata value for dimensions that have
been set.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>