[SYCL] Improve ODS negative filter implementation (#7453)
The negative filter implementation for ONEAPI_DEVICE_SELECTOR uses a map
to keep track of blacklisted devices. The keys used by this map were
originally device addresses in a vector container which are not very
robust because vectors can potentially move their data to other
locations and the device addresses could change thus invalidating the
blacklist map. Even though in the source code the resizing of the vector
only happens after we are done with the blacklist, you never know what
tricks the compiler might pull on us. We use device numbers instead
which are unique for each device in a platform and do not change during
the function execution.