[SYCL] [CI] Add OCL dev env to containers. (#7440)
Adding OpenCL dev env to our containers. At minimum, this will allow our
containers to be used for interop testing on GitHub Actions (which is
being skipped presently).

L0 drivers are present on the "drivers" image and don't need to be
installed. Plus the L0 interop tests are not being skipped.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>

Signed-off-by: Chris Perkins <chris.perkins@intel.com>