[SYCL] Fix ModuleSplitterBase::totalSplits usage
`ModuleSplitterBase::totalSplits` method returns the amount of
_remaining_ splits, i.e. every call to `ModuleSplitterBase::nextSplit`
decrements that value.

Fixed an incorrect initialization of `SplitByMode` and
`SplitByOptionalFeatures` variables by re-arranging code so we ask for
amount of splits _before_ we extract them. Previously, those two
variables were always initialized with `false`: from what I see in a
code this shouldn't have resulted in any major issues, but it caused a
problem in a downstream.