Modified post link test to reintroduce compile time properties
sycl-post-link splits the test into three translation units, if the pass
is run ahead of time, the annotations and properties aren't split properly
so the pass is run after the split per TU to keep the test working.

If the pass can be run before the split, this test can be properly adapted,
but this is the only way I can think to preserve this test and maintain
coverage as is for now.