[SYCL][Fusion] Add kernel fusion extension API (#7416)
This is the first patch in a series of patches to add an implementation
of the [kernel fusion
extension](https://github.com/intel/llvm/pull/7098). We have split the
implementation into multiple patches to make them more easy to review.

This first patch introduces the user-facing API discussed in the
[extension proposal](https://github.com/intel/llvm/pull/7098). It does
not yet add any fusion functionality, just the mere API and SYCL
properties. Calls to the API will throw an error until the remaining
functionality lands with the following patches.

The design document for the overall implementation of kernel fusion can
be found [here](https://github.com/intel/llvm/pull/7204).

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>