Driver changes to implement -fsycl-no-rdc
Today, the basic flow of the SYCL device-side driver is as follows

1) Link all device code from all compiler inputs together
2) Link 1) against SYCL device libraries
3) Run sycl-post-link on 2)
4) Run llvm-spirv on 3)
5) Run clang-offload-wrapper on 4)

Step 1 can create a performance bottleneck when you have a huge number of kernels.
If none of the kernels use globals or SYCL_EXTERNAL functions, we can actually split it up to be the following:

For object file:
   1) Link this object file's device code with SYCL device libraries
   2) Run sycl-post-link on 1)
   3) Run llvm-spirv on 2)
   4) Run clang-offload-wrapper on 4)

Since we don't link all device code together, each step runs on smaller IR which results in
compiler runtime and compiler memory usage benefits.

Note that in order to do the above per-object-file, we need to break up fat static arhives. We do
this by using the ForEachWrappingAction action. This allows us to run commands on each item inside the
fat static archive

The driver flow when a static archive is involved is the most complex case and looks like the below:

1) spriv-to-ir-wrapper on fat static archive
2) Link all SYCL device libraries together without any user device code into a single device library BC
3) llvm-foreach:
  3a): llvm-link current object file with 2)
4) file-table-tform replacing tempfilelist from 1) with output of 3)
5) llvm-foreach:
  5a) sycl-post-link on 4)
6) llvm-foreach:
  6a) Extract BC file column from 5) output table
7) file-table-tform on 6) merging all BC file columns into a single big column
8) llvm-spirv on 7) (does llvm-foreach internally)
9) file-table-tform on 5) output table, replacing BC column with spirv column from 8)
10) clang-offload-wrapper with 9)

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>