[LIBCLC] Generate FunctionDecl to remangle substituted entries (#7220)
Retire the by hand mangling of substituted function in favor of
constructing corresponding `FunctionDecls` and letting clang mangle it
for us. This ensures that we never diverge from clang's mangling
idiosyncrasies and bugfixes would likely be in the AST creation, not
mangling itself.

The tool has been rewritten as a `ClangTool` implementing
`FrontendAction`.

Additionally, we provide the option to perform a test run, in which no
substitutions are made, while checking that remangled name does not
diverge from the original mangled name. The tests are added to
`check-libclc` target.

The diffs are a bit ugly, it might be easier to inspect the file mode,
basically the change lives entirely in `Remangler` class, and if you
follow `Remangler::remangle` it will lead you through what it does.

Fixes: https://github.com/intel/llvm/issues/6505

Co-authored-by: Aaron Ballman <aaron@aaronballman.com>
Co-authored-by: premanandrao <premanand.m.rao@intel.com>
Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>