Reorder invocations of fixType, ReplaceTypeInOperands and mergeType
The reason of change is to mitigate Coverity warnings from CMPLRLLVM-42088.
The issue is that SPIRVTypeScavenger::fixType method
performs `delete &Ty` at the end. Then this freed
pointer is used later just as value (not dereferenced).
Nevertheles, Coverity claims use-after-free warning.
This commit reorders function's invocations to mitigate this issue.

Probably, a better approach could be a move to "C++ style"
memory management.