Optimization pipeline refactor to honour flags and fix logic
All of the SYCL passes now are no longer run when `-disable-llvm-passes`
is set, so we now properly honour the flag in this case.
Several passes also run in under correct circumstances now.

This has caused the following tests to fail:
- CodeGenSYCL/group-local-memory.cpp
- SemaSYCL/sycl-force-inline-kernel-lambda.cpp
- CodeGenSYCL/device_has.cpp
- CodeGenSYCL/sub-group-size.cpp
- CodeGenSYCL/uses_aspects.cpp

These all be addressed in the next few commits, or in some cases in new
PRs due to scope and needing discussion, etc.