[SYCL] Start events cleanup for immediate command lists based on threshold (#7773)
For regular command lists we check the status of the command list fence
quite often - every time when command list is requested and there is no
available command list. When fence is completed we cleanup events from
the command list.

Currently for immediate command lists events get cleaned up only at
points of syncrhonization which may be too rare for some applications.
So, make the cleanup process to be started if number of events in the
immediate command list exceeds threshold. Also introduce a switch which
allows to control/tune.

This results in better recycling of events for immediate command lists
and improves performance.