[SYCL][Docs] Clarifies default device_global allocation behavior
Currently the device_global extension specifies that the same allocation
used by a device_global is accessible by other kernels on the same
device, however memory allocations for devices are not guaranteed to be
accessible across contexts. As such, this commit clarifies that the
consistency between kernels is only true for kernels running on the same
device and context.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>