[SYCL] Use decorated pointer in device_global
device_global without device_image_scope currently use an undecorated
pointer for its underlying type, but we know that the pointer will be
to the global memory space. This commit changes the underlying member
to be a decorated pointer and change the get_multi_ptr member function
to create the multi_ptr directly from the decorated pointer instead of
performing unnecessary address space casts.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>