[SYCL][sycl-post-link] Generate global_id_mapping for device_global name mapping
Some backends do not natively support the HostAccessINTEL decoration and
instead need to know the name of the device_global in the device
program. However, the SYCL runtime will only be informed about the
unique name, so this commit adds program metadata for making the mapping
between unique names and actual names in the programs, allowing the
backends to make the mappings internally.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>