[CI] Add CUDA on AWS run in pre-commit
Reintroduce the change #7790

Add CUDA on AWS in addition to CUDA on self-hosted runner. Self-hosted
runner will be turned off for OS upgrade.
Note that pre-commit testing on this PR will test nothing due to
pull_request_target trigger.
This change has to be merged first and we'll see effect on the other
pre-commit PRs where we expect both self-hosted and AWS run done in
parallel.

Tested on #7806