Converted inlining test to check nodes in AST
The -fsycl-force-inline-kernel-lambda flag adds the inlining attribute to the
AST node of the kernel, so it was necessary to make this test more specific.

The attributes aren't put on the FunctionDecl node of each class template argument
to parallel_for, but rather on the actual LambdaExpr subtree it calls.

Maybe some additional CHECKs to link the LambdaExpr and FunctionDecl of
the kernel would be good future work, as the kernel FunctionalDecl aren't in
the same location as the Lambda in which they're called. There is a reference
linking the two together, but as far as I can tell, no functionality to link these
together, as the curent CHECK functions seem to regex match individual lines?