Refactored pipeline building logic after removing first SYCLPropagateAspectsUsagePass
The first registration of the pass was a holdover from when
-disable-llvm-passes was also bound to -fno-sycl-early-optimizations and
the pass needed to be run in a particular instance. Now that this is no
longer the case, this extra call has been removed.

This did cause 2 test failures though, as the main invocation of this pass
wasn't being run when the flag was present as the pipeline building logic
was only invoked if it wasn't present. Now a O0 pipeline is built if the
flag is set, as a fallback.