[SYCL][ESIMD] Add semi-dynamic SLM allocation - esimd::experimental::slm_allocator. (#7759)
This patch adds new class - slm_allocator - and its lowering. This is
RAII-style class used to implement "semi-dynamic" SLM allocation. SLM is
allocated in the constructor and released in the destructor, that's why
it is "dynamic", as opposed to fully static allocation style of
'slm_init'. Actual offset of SLM chunk allocated by the call is
calculated at compile time, that's why it is "semi-". To calculate SLM
usage by a kernel, compiler finds a path in a callgraph with the largest
amount of SLM "locked" by slm_allocator objects live along the paths.
slm_init call also participates in calculating SLM budget. It can be
modelled as slm_allocator object declared at the very beginning of a
kernel and live till its the very end.

Since a call graph is used, function pointers and recursion is not
supported.
Complementary E2E test:
https://github.com/intel/llvm-test-suite/pull/1449.

Signed-off-by: Konstantin S Bobrovsky <konstantin.s.bobrovsky@intel.com>