[SYCL][NFCI] Move default optimization level configuration
Right now the front-end compiler modifies default optimization level in
SYCL device compilation mode (upstream clang uses O0, whereas DPC++ uses
O2). As some of CodeGen module patches landed to upstream, this
difference causes conflicts and regresions when we pull changes from the
upstream to the `sycl` branch.

Current optimization level configuration is modeled after OpenCL
compiler, which is a JIT compiler, so it doesn't use clang Driver.
For DPC++ compiler it's more reasonable to make this type of changes in
the driver to avoid significant divergence in LIT tests.