[SYCL][XPTI] Fix building with XPTI disabled
The `XPTIScope` class uses some XPTI types, but the XPTI header is only
included if the XPTI tracing is enabled, so when the XPTI tracing is
disabled it fails to build.

So this patch simply macros out the whole `XPTIScope` class rather than
specific member functions.

All the uses of `XPTIScope` in the codebase are already guarded by the
macro.