[SYCL][Doc] Update doc for No RDC mode (#7813)
Update doc for No RDC mode.

I considered updating the diagrams or adding new ones, but the current
ones are very simplified, contain no conditional logic and already skip
cases today. We would need complex diagrams with conditional logic to
accurately capture this feature which would become a barrier for users
understanding the driver flow in most cases. An accurate diagram is
available in the source code as a comment
[here](https://github.com/intel/llvm/blob/sycl/clang/lib/Driver/Driver.cpp#L5435).


Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>