[ESIMD][NFC] Fix warning: unused variable StoreInstValueOperandIndex (#7912)
The variable was used only in 1 assert statement and caused a warning
with build modes where assert is mapped to NO-OP.

The assert is also removed because it verifies something that should
always be true. In particular, the StoreInst instructions have 2
operands: Value and DestPtr. If 'StI' is user of 'U' and 'U' is not
DstPtr, then 'U' simply must be the 'Value' operand.

Signed-off-by: Vyacheslav N Klochkov <vyacheslav.n.klochkov@intel.com>