[SYCL] Reimplemented -f[no]sycl-early-optimizations flag (#7701)
Reimplemented the `-f[no]sycl-early-optimizations` compiler flag to
separate it from the meaning of `-disable-llvm-passes` for more
fidelity. This required a change to its definition, setting of a new
codegen option behind-the-scenes, and small logic changes to the
optimization pipeline to factor in the new flag. Existing tests all
still pass.