[SYCL] Add RegisterAllocMode metadata to large GRF kernels
Currently, we tell IGC to compile the kernel in large GRF mode
by the -doubleGRF/-ze-opt-large-register-file flags passed
in program manager. However, this doesn't work for AOT
because the kernel has already been compiled and today we have no way
to change AOT compiler flags based on the kernel contents. Instead,
use the RegisterAllocMode metadata to get IGC to use large
GRF mode, which will work for both JIT and AOT. This metadata
is translated to the final metadata IGC looks for in llvm-spirv.

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>