[SYCL][CUDA] Allow FTZ, prec-sqrt to override no-ftz, no-prec-sqrt (#7616)
If two bc files are compiled with different values for flags
`nvvm-reflect-ftz` or `nvvm-reflect-prec-sqrt` then llvm-link will emit
an error for the conflicting module flags. This instead allows FTZ=true
to override FTZ=false, and the same with `prec-sqrt`.

This was blocking ftz=true from being used in SYCL for CUDA backend as
the `llvm/libdevice` library is compiled with default ftz value, meaning
introducing a non default value for `nvvm-reflect-ftz` will fail at
`llvm-link` time.

An alternative is to introduce a clang flag that will completely omit
nvvm-reflect module flags, which can then be used when compiling
libdevice.